moved plot_template to settings folder
