+++ protzilla/data_preprocessing/plots.py
@@ -5,9 +5,9 @@ import plotly.graph_objects as go
 from plotly.graph_objects import Figure
 from plotly.subplots import make_subplots
 
+from ui.settings.plot_template import *
 from protzilla.data_preprocessing.plots_helper import generate_tics
 from protzilla.utilities import default_intensity_column
-from protzilla.utilities.plot_template import *
 from protzilla.constants.colors import *
 
 def create_pie_plot(
