+++ ui/settings/plot_template.py
